streamlit/Routine/requirements.txt
altair==5.5.0
altgraph==0.17.4
anyio==4.9.0
arrow==1.3.0
asttokens==3.0.0
attrs==25.3.0
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
cookiecutter==2.6.0
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
flet==0.27.6
flet-cli==0.27.6
flet-desktop==0.27.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keyboard==0.13.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
MouseInfo==0.1.3
msal==1.32.3
narwhals==1.35.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.2.2
Office365-REST-Python-Client==2.6.1
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.7
plyer==2.1.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.4.0
qrcode==7.4.2
referencing==0.36.2
repath==0.9.0
requests==2.32.3
requests-negotiate-sspi==0.5.2
requests_ntlm==1.3.0
rich==14.0.0
rpds-py==0.24.0
selenium==4.32.0
setuptools==80.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sspilib==0.3.0
st-annotated-text==4.0.2
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
